requirements.txt
tqdm==4.65.0
notebook==7.1.2
groq
elasticsearch==8.13.0
pandas==1.5.3
scikit-learn==1.2.2
ipywidgets==8.0.6
minsearch
matplotlib==3.7.1
seaborn==0.12.2
numpy==1.24.3
python-dotenv==1.0.0
sentence_transformers==2.2.2
flask==2.3.2
requests==2.31.0